--- src/app/features/public/components/donations/donations.component.html
@@ -1,38 +1,0 @@
-    <alk-donations-header [message]="message" ></alk-donations-header>
-    <section class="Donaciones-HaceDonacion  py-5">
-        <div class="container py-3">
-          <div class="row justify-content-center">
-      
-            <div class="col col-12 col-sm-12 col-md-10 col-lg-12 mb-4 module-one_card option-selection-card active-card">
-              <div class="BoxNewDonation Bg-White w-75 p-3 p-lg-5 pb-5 m-auto">
-                <form [formGroup]="form"  (ngSubmit)="donation()">
-                <h4 class="Title BoxTitleDonation text-center Cl-GrisOscuro mb-4">Hacé tu donación ahora</h4>
-      
-                <div class="BoxTypeDonation BoxTypeDonation-Mensual mt-4 mt-lg-5 option-selection option-selection-module-mensual active">
-                  <div>
-                    <a (click)="donar1()"  class="BtnDonationImport BtnDefault">$500</a>
-                    <a (click)="donar2()"  class="BtnDonationImport BtnDefault">$1000</a>
-                    <a (click)="donar3()"  class="BtnDonationImport BtnDefault">$1500</a>
-                    <a (click)="donar4()"  class="BtnDonationImport BtnDefault">$2000</a>
-                    <a (click)="donar5()"  class="BtnDonationImport BtnDefault">$2500</a>
-                    <a (click)="donar6()"  class="BtnDonationImport BtnDefault">Otro Importe</a>
-                  </div>
-
-                  <input type="hidden" class="form-control mb-4"  formControlName="value"  id="value"  name="value" placeholder="Ingrese el importe y luego haga click en el boton Donar." >
-                  <div>
-                    <span appMsgErrorForm messageError="Por favor, ingrese un valor" *ngIf="empty"></span> <br />
-                    <span class="invalid" *ngIf="form.get('value')?.errors?.['required'] &&  form.get('value')?.touched"
-                      >El importe de la donacion es requerido
-                    </span>
-                    <span appMsgErrorForm messageError="Por favor, ingrese un valor" *ngIf="empty">Debe completar el importe a donar para continuar.</span> <br>
-                    <span appMsgErrorForm messageError="Por favor, ingrese un valor" *ngIf="error">Los datos ingresados son invalidos.</span> <br>
-                  </div>
-                  <button class="w-50 BtnDefault mt-4 mb-4 m-auto" type="submit">Donar</button>
-                </div>
-            </form>
-              </div>
-            </div>
-            
-          </div>
-        </div>
-      </section>
