--- src/app/features/public/components/donations/donations.component.scss
@@ -1,333 +1,0 @@
-section {
-  display: block;
-}
-h4 {
-  margin-top: 0;
-  margin-bottom: 0.5rem;
-}
-p {
-  margin-top: 0;
-  margin-bottom: 1rem;
-}
-select {
-  text-transform: none;
-}
-h4 {
-  margin-bottom: 0.5rem;
-  font-family: inherit;
-  font-weight: 500;
-  line-height: 1.2;
-  color: inherit;
-}
-h4 {
-  font-size: 1.5rem;
-}
-.container {
-  width: 100%;
-  padding-right: 15px;
-  padding-left: 15px;
-  margin-right: auto;
-  margin-left: auto;
-}
-@media (min-width: 576px) {
-  .container {
-    max-width: 540px;
-  }
-}
-@media (min-width: 768px) {
-  .container {
-    max-width: 720px;
-  }
-}
-@media (min-width: 992px) {
-  .container {
-    max-width: 960px;
-  }
-}
-@media (min-width: 1200px) {
-  .container {
-    max-width: 1140px;
-  }
-}
-.row {
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  -ms-flex-wrap: wrap;
-  flex-wrap: wrap;
-  margin-right: -15px;
-  margin-left: -15px;
-}
-.col,
-.col-12,
-.col-lg-12,
-.col-lg-6,
-.col-md-10,
-.col-md-6,
-.col-sm-12,
-.col-sm-6 {
-  position: relative;
-  width: 100%;
-  min-height: 1px;
-  padding-right: 15px;
-  padding-left: 15px;
-}
-.col {
-  -ms-flex-preferred-size: 0;
-  flex-basis: 0;
-  -webkit-box-flex: 1;
-  -ms-flex-positive: 1;
-  flex-grow: 1;
-  max-width: 100%;
-}
-.col-12 {
-  -webkit-box-flex: 0;
-  -ms-flex: 0 0 100%;
-  flex: 0 0 100%;
-  max-width: 100%;
-}
-@media (min-width: 576px) {
-  .col-sm-6 {
-    -webkit-box-flex: 0;
-    -ms-flex: 0 0 50%;
-    flex: 0 0 50%;
-    max-width: 50%;
-  }
-  .col-sm-12 {
-    -webkit-box-flex: 0;
-    -ms-flex: 0 0 100%;
-    flex: 0 0 100%;
-    max-width: 100%;
-  }
-}
-@media (min-width: 768px) {
-  .col-md-6 {
-    -webkit-box-flex: 0;
-    -ms-flex: 0 0 50%;
-    flex: 0 0 50%;
-    max-width: 50%;
-  }
-  .col-md-10 {
-    -webkit-box-flex: 0;
-    -ms-flex: 0 0 83.333333%;
-    flex: 0 0 83.333333%;
-    max-width: 83.333333%;
-  }
-}
-@media (min-width: 992px) {
-  .col-lg-6 {
-    -webkit-box-flex: 0;
-    -ms-flex: 0 0 50%;
-    flex: 0 0 50%;
-    max-width: 50%;
-  }
-  .col-lg-12 {
-    -webkit-box-flex: 0;
-    -ms-flex: 0 0 100%;
-    flex: 0 0 100%;
-    max-width: 100%;
-  }
-}
-.form-control {
-  display: block;
-  width: 100%;
-  padding: 0.375rem 0.75rem;
-  font-size: 1rem;
-  line-height: 1.5;
-  color: #495057;
-  background-color: #fff;
-  background-clip: padding-box;
-  border: 1px solid #ced4da;
-  border-radius: 0.25rem;
-  transition: border-color 0.15s ease-in-out, box-shadow 0.15s ease-in-out;
-}
-.form-control::-ms-expand {
-  background-color: transparent;
-  border: 0;
-}
-.form-control:focus {
-  color: #495057;
-  background-color: #fff;
-  border-color: #80bdff;
-  outline: 0;
-  box-shadow: 0 0 0 0.2rem rgba(0, 123, 255, 0.25);
-}
-.form-control:focus {
-  color: #495057;
-  background-color: #fff;
-  border-color: #80bdff;
-  outline: 0;
-  box-shadow: 0 0 0 0.2rem rgba(0, 123, 255, 0.25);
-}
-.form-control::-webkit-input-placeholder {
-  color: #6c757d;
-  opacity: 1;
-}
-.form-control::-moz-placeholder {
-  color: #6c757d;
-  opacity: 1;
-}
-.form-control:-ms-input-placeholder {
-  color: #6c757d;
-  opacity: 1;
-}
-.form-control::-ms-input-placeholder {
-  color: #6c757d;
-  opacity: 1;
-}
-.form-control::placeholder {
-  color: #6c757d;
-  opacity: 1;
-}
-.form-control:disabled {
-  background-color: #e9ecef;
-  opacity: 1;
-}
-select.form-control:not([size]):not([multiple]) {
-  height: calc(2.25rem + 2px);
-}
-.form-group {
-  margin-bottom: 1rem;
-}
-.justify-content-center {
-  -webkit-box-pack: center !important;
-  -ms-flex-pack: center !important;
-  justify-content: center !important;
-}
-.w-50 {
-  width: 50% !important;
-}
-.w-75 {
-  width: 75% !important;
-}
-.mt-4 {
-  margin-top: 1.5rem !important;
-}
-.p-3 {
-  padding: 1rem !important;
-}
-.py-3 {
-  padding-top: 1rem !important;
-}
-.py-3 {
-  padding-bottom: 1rem !important;
-}
-.py-5 {
-  padding-top: 3rem !important;
-}
-.pb-5,
-.py-5 {
-  padding-bottom: 3rem !important;
-}
-.m-auto {
-  margin: auto !important;
-}
-@media (min-width: 992px) {
-  .mt-lg-5 {
-    margin-top: 3rem !important;
-  }
-  .p-lg-5 {
-    padding: 3rem !important;
-  }
-  .py-lg-5 {
-    padding-top: 3rem !important;
-  }
-  .pr-lg-5 {
-    padding-right: 3rem !important;
-  }
-  .py-lg-5 {
-    padding-bottom: 3rem !important;
-  }
-}
-.text-center {
-  text-align: center !important;
-}
-
-.Bg-White {
-  background: #aebac9 !important;
-}
-.Cl-GrisOscuro {
-  color: #404041 !important;
-}
-.BoxTitleDonation {
-  font-size: 2rem !important;
-  line-height: 2.2rem !important;
-  font-weight: 700 !important;
-}
-.BtnDefault {
-  background: #ffce35;
-  color: #404041;
-  border-radius: 50px;
-  padding: 12px 36px;
-  border: 1px solid #ffce35;
-  text-transform: uppercase;
-  text-align: center;
-  font-weight: 700;
-  display: inline-block;
-  text-decoration: none;
-  box-shadow: 0px 12px 15px rgba(0, 0, 0, 0.2);
-  min-width: 220px;
-  font-size: 0.9rem;
-  max-height: 48px;
-  z-index: 1;
-  cursor: pointer;
-}
-.BtnDefault:hover {
-  background: #404041;
-  color: #fff;
-  text-decoration: none;
-  border: 1px solid #404041;
-}
-.BoxNewDonation {
-  box-shadow: 0px 12px 15px rgba(0, 0, 0, 0.2);
-  border-radius: 10px;
-  display: flex;
-  flex-direction: column;
-}
-.BoxTypeDonation {
-  display: flex;
-  justify-content: space-between;
-  flex-wrap: wrap;
-}
-.BoxTypeDonation .BtnDonationImport {
-  background: #eae5dc;
-  border-color: #eae5dc;
-  margin-bottom: 2rem;
-  box-shadow: none;
-}
-.BoxTypeDonation .BtnDonationImport.Active,
-.BoxTypeDonation .BtnDonationImport:hover {
-  background: #7bbfb3;
-  border-color: #7bbfb3;
-  color: #fff;
-}
-@media (max-width: 990px) {
-  .BoxNewDonation {
-    width: 100% !important;
-  }
-  .BoxTypeDonation .BtnDonationImport {
-    margin-bottom: 0.5rem;
-  }
-  .BoxTitleDonation {
-    font-size: 1.6rem !important;
-    line-height: 1.8rem !important;
-    color: white;
-  }
-}
-@media (max-width: 767px) {
-  .BtnDefault {
-    min-width: 100%;
-  }
-}
-.BoxTypeDonation {
-  display: none;
-}
-.BoxTypeDonation > div {
-  width: 100%;
-  display: flex;
-  justify-content: space-between;
-  flex-wrap: wrap;
-}
-.BoxTypeDonation.active {
-  display: flex;
-}
