+++ src/app/features/public/pages/donations/donations-header/donations-header.component.scss
@@ -41,7 +41,7 @@
 
 .banner{
     height: 700px;
-    background: url("../../../../../../assets/img/public/donations/donations-banner.jpg") fixed center/cover;
+    background: url("/assets/public/donations/donations-banner.jpg") fixed center/cover;
     display: flex;
     align-items: center;
     justify-content: center;
